+++ src/components/Field/CHANGELOG.md
@@ -1,1 +1,2 @@
 - 2025-10-10: Initial version.
+- 2025-10-12: Add `Field.Select` component.